# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water traker app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of a water tracker app"
  },
  "servers": [
    {
      "url": "https://tracker-of-water-xk7t.onrender.com"
    }
  ],
...
  "paths": {
...
    "/water/{id}": {
      "patch": {
        "tags": [
          "Water"
        ],
        "summary": "Edit water consumption record",
        "operationId": "patchWater",
        "description": "Update an existing water consumption record by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id of the water consumption record to update",
            "required": true,
            "schema": {
              "type": "string",
              "example": "20b2f76d822221c0b0f394a1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "waterVolume",
                  "data"
                ],
                "properties": {
                  "waterVolume": {
                    "type": "number",
                    "description": "waterVolume",
                    "example": 500
                  },
                  "data": {
                    "type": "string",
                    "description": "data`s water",
                    "example": "2024-10-23 13:10"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Water record updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Water record updated successfully!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "Water"
        ],
        "summary": "Delete water consumption record",
        "operationId": "deleteWater",
        "description": "Delete an existing water consumption record by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The id of the water consumption record to delete",
            "required": true,
            "schema": {
              "type": "string",
              "example": "20b2f76d822221c0b0f394a1"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Water consumption record deleted successfully. No content returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 204
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully deleted water"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "water": {
        "type": "object",
        "required": [
          "waterVolume",
          "date"
        ],
        "properties": {
          "waterVolume": {
            "type": "number",
            "description": "The waterVolume of water consumed, in milliliters (must be between 50 and 5000).",
            "minimum": 1,
            "maximum": 5000,
            "example": 250
          },
          "userId": {
            "type": "string",
            "description": "User's ID",
            "example": "609e126aa08b5b3c88b728db"
          },
          "date": {
            "type": "string",
            "description": "The date and time of water consumption. Must be in 'DD-MM-YYYY HH:mm' format",
            "format": "date-time",
            "example": "17-12-2024 18:26"
          },
          "_id": {
            "type": "string",
            "example": "67135ea1260aecfe82bea7b7"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-17T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-17T12:00:00.000Z"
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "404": {
        "description": "Contact not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Contact not found"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}